chore: reogranize some stuff
